Split epoch_pds into submodules, keep tests in a different file but still as a pds/ submodule
