+++ src/pds/mod.rs
@@ -1,1 +1,6 @@
+pub mod accounting;
 pub mod epoch_pds;
+pub mod quotas;
+
+#[cfg(test)]
+mod tests;
