+++ src/queries/histogram.rs
@@ -53,10 +53,15 @@ where
     /// will be obtained by retrieving *relevant* events from the event
     /// storage.
     ///
-    /// Events with value 0 can be omitted since whey won't appear in the `compute_report` sum. Histogram buckets are padded with zeros at aggregation time.
+    /// Events with value 0 can be omitted since whey won't appear in the
+    /// `compute_report` sum. Histogram buckets are padded with zeros at
+    /// aggregation time.
     ///
     /// Events can point to the relevant_events_per_epoch, hence
-    /// the lifetime. Returns an (ordered) vector of tuples (event, value), which will be browsed in order by `compute_histogram_report`. Ordering can depend on timestamp or other conditions (e.g. order of the vector of events for one epoch) in the implementation.
+    /// the lifetime. Returns an (ordered) vector of tuples (event, value),
+    /// which will be browsed in order by `compute_histogram_report`. Ordering
+    /// can depend on timestamp or other conditions (e.g. order of the vector of
+    /// events for one epoch) in the implementation.
     fn event_values<'a>(
         &self,
         relevant_events_per_epoch: &'a HashMap<
@@ -102,8 +107,8 @@ where
         // selected, so the sum is juat that singular value.
         //
         // The order matters, since events that are attributed last might be
-        // dropped by the contribution cap. `event_values` is in charge of ordering
-        // the events from `relevant_events_per_epoch`.
+        // dropped by the contribution cap. `event_values` is in charge of
+        // ordering the events from `relevant_events_per_epoch`.
         let mut report = HistogramReport {
             bin_values: HashMap::new(),
         };
